funciona

+++ 2 - Ising/Códigos/Ising/funcion_flip.h
@@ -1,4 +1,4 @@
 #ifndef funcion_flip_H
 #define funcion_flip_H
-int flip(double *red,double *tabla,int N,int L);
+int flip(int *red,double *tabla,int N,int L);
 #endif
