+++ 2 - Ising/Códigos/Ising/funcion_imprimir.h
@@ -1,4 +1,4 @@
 #ifndef funcion_imprimir_H
 #define funcion_imprimir_H
-int imprimir(double *red, int L);
+int imprimir(int *red, int L);
 #endif
