+++ 2 - Ising/Códigos/Ising/funcion_poblar.h
@@ -1,4 +1,4 @@
 #ifndef funcion_poblar_H
 #define funcion_poblar_H
-int poblar(float p, int L,double *red);
+int poblar(float p, int L,int *red);
 #endif
